[#20] enum BookType Constants 폴더로 변경
